Adapt access_through_char_pointer.c for c6502's type widths
The test's premise — "inspect the four bytes of an int" — assumes
sizeof(int) == 4, which holds upstream but not on c6502 (int is
2 bytes). Switch the local from `int` to `long` (4B in c6502) so
the four-byte readout actually has four bytes to read.

Same fix for the array probe: `unsigned int array[3][2][1]` becomes
`unsigned long array[3][2][1]` so each two-element row really is 8
bytes (matching the test's `+16` skip-two-rows offset). The
0xFFFF0000 literal needs a `ul` suffix instead of `u` because
unsigned int (2B) maxes at 0xFFFF.

Plain `char` is unsigned in c6502, but the test compares bytes
against negative literals (-128, -1) on the assumption that plain
char is signed. Use `signed char *` for the byte pointer so those
comparisons see signed bytes regardless of the implementation
choice for plain char.

Removed from test_sim_asm SKIPS (now passes through asm sim) and
flipped the EXPECTED_RETURNS in test_sim_chapters from 3 → 0.

+++ tests/chapter_16/valid/chars/access_through_char_pointer.c
@@ -5,9 +5,13 @@
 
 int main(void) {
 
-    // inspect the four bytes of an int
-    int x = 100;
-    char *byte_ptr = (char *) &x;
+    // inspect the four bytes of a long (in c6502, `long` is 4 bytes
+    // — `int` is only 2, so it can't satisfy the test's premise).
+    // Use `signed char *` for the byte pointer because plain `char`
+    // is unsigned in c6502, and the test compares bytes against
+    // negative values like -128 / -1.
+    long x = 100;
+    signed char *byte_ptr = (signed char *) &x;
 
     if (byte_ptr[0] != 100) {
         return 1;
@@ -19,7 +23,7 @@ int main(void) {
 
     // now inspect a double -- only upper bit should be set
     double d = -0.0; // 0x8000_0000_0000_0000
-    byte_ptr = (char *) &d;
+    byte_ptr = (signed char *) &d;
     if (byte_ptr[7] != -128) {
         return 3;
     }
@@ -31,13 +35,13 @@ int main(void) {
     }
 
     // finally, let's look at an array
-    unsigned int array[3][2][1] = {
+    unsigned long array[3][2][1] = {
         {{-1}, {-1}},
         {{-1}, {-1}},
-        {{4294901760u}} // 0xffff_0000
+        {{4294901760ul}} // 0xffff_0000
     };
-    byte_ptr = (char *) array;
-    byte_ptr = byte_ptr + 16; // each row is 8 bytes since it has 2 ints
+    byte_ptr = (signed char *) array;
+    byte_ptr = byte_ptr + 16; // each row is 8 bytes since it has 2 longs
     if (byte_ptr[0] || byte_ptr[1]) {
         return 5;
     }
